% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator_msms.R
\name{activate_msms}
\alias{activate_msms}
\title{Use the simulator msms}
\usage{
activate_msms(jar, java, priority = 200)
}
\arguments{
\item{jar}{The path of the msms jar file.}

\item{java}{The path of the java executable on your system.}

\item{priority}{The priority for this simulator. If multiple simulators
can simulate a model, the one with the highest priority will be used.}
}
\description{
This adds the simulator 'msms' to the list of available simulators. To add
msms, you need to download the jar file and have Java installed on your
system.
}
\section{Citation}{

Gregory Ewing and Joachim Hermisson.
MSMS: a coalescent simulation program including recombination,
demographic structure and selection at a single locus.
Bioinformatics (2010) 26 (16): 2064-2065
doi:10.1093/bioinformatics/btq322
}

