% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator_seqgen.R
\name{activate_seqgen}
\alias{activate_seqgen}
\title{Use seq-gen for simulating finite sites mutation models}
\usage{
activate_seqgen(binary, priority = 100)
}
\arguments{
\item{binary}{The path of the seqgen binary that will be used
for simulations.}

\item{priority}{The priority for this simulator. If multiple simulators
can simulate a model, the one with the highest priority will be used.}
}
\description{
This allows you to use seq-gen to simulate finite sites mutation models.
You need to download the program from
\url{http://tree.bio.ed.ac.uk/software/seqgen/}
and compile the binary first.
}
\section{Citation}{

Andrew Rambaut and Nicholas C. Grass.
Seq-Gen: an application for the Monte Carlo simulation of DNA sequence
evolution along phylogenetic trees.
Comput Appl Biosci (1997) 13 (3): 235-238
doi:10.1093/bioinformatics/13.3.235
}

