% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsites.R
\name{as.segsites}
\alias{as.segsites}
\title{Convert genetic data to coala's internal format}
\usage{
as.segsites(data, ...)
}
\arguments{
\item{data}{The data object that is converted.}

\item{...}{Additional arguments specific for the used package.}
}
\value{
A list of segregating sites objects.
}
\description{
This function can be used to convert the genomic data formats used in other
packages to calculate coala's segregating sites object. This is useful for
calculating the summary statistics of biological data using the
\code{\link{calc_sumstats_from_data}} function.
}
\details{
Currently, only the package \pkg{PopGenome} is supported, see
\code{\link{as.segsites.GENOME}} for details.
}

