% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{check_model}
\alias{check_model}
\title{Check which simulator can simulate a model}
\usage{
check_model(model)
}
\arguments{
\item{model}{The model that is checked}
}
\description{
This function allows to check which of the available simulators can
simulate a given model. It also states the problems for the ones that
are incompatible with the model.
}

