% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coala.R
\docType{package}
\name{coala-package}
\alias{coala}
\alias{coala-package}
\title{A Framework for Coalescent Simulation in R}
\description{
This package allows to specify and simulate coalescent models from
within R. The `introduction` vignettes is a good place to start.
}

