% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coalsimr.R
\docType{package}
\name{coalsimr-package}
\alias{coalsimr-package}
\title{Fast estimation of demographic parameters}
\description{
Fast estimation of demographic parameters
}
\author{
Paul R. Staab \email{develop (at) paulstaab.de}
}
\keyword{package}

