% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_growth.R, R/feature_migration.R, R/feature_mutation.R, R/feature_outgroup.R, R/feature_pop_merge.R, R/feature_recombination.R, R/feature_sample.R, R/feature_selection.R, R/feature_size_change.R, R/feature_sumstats.R, R/feature_unphased.R, R/simulator_ms.R, R/simulator_msms.R, R/simulator_scrm.R, R/simulator_seqgen.R
\name{conv_to_ms_arg.Feature_growth}
\alias{conv_to_ms_arg}
\alias{conv_to_ms_arg.Feature_files}
\alias{conv_to_ms_arg.Feature_growth}
\alias{conv_to_ms_arg.Feature_migration}
\alias{conv_to_ms_arg.Feature_mutation}
\alias{conv_to_ms_arg.Feature_outgroup}
\alias{conv_to_ms_arg.Feature_pop_merge}
\alias{conv_to_ms_arg.Feature_recombination}
\alias{conv_to_ms_arg.Feature_sample}
\alias{conv_to_ms_arg.Feature_seg_sites}
\alias{conv_to_ms_arg.Feature_selection}
\alias{conv_to_ms_arg.Feature_size_change}
\alias{conv_to_ms_arg.Feature_trees}
\alias{conv_to_ms_arg.Feature_unphased}
\alias{conv_to_ms_arg.default}
\alias{conv_to_msms_arg.Feature_files}
\alias{conv_to_msms_arg.Feature_growth}
\alias{conv_to_msms_arg.Feature_migration}
\alias{conv_to_msms_arg.Feature_mutation}
\alias{conv_to_msms_arg.Feature_outgroup}
\alias{conv_to_msms_arg.Feature_pop_merge}
\alias{conv_to_msms_arg.Feature_recombination}
\alias{conv_to_msms_arg.Feature_sample}
\alias{conv_to_msms_arg.Feature_seg_sites}
\alias{conv_to_msms_arg.Feature_selection}
\alias{conv_to_msms_arg.Feature_size_change}
\alias{conv_to_msms_arg.Feature_trees}
\alias{conv_to_msms_arg.Feature_unphased}
\alias{conv_to_msms_arg.default}
\alias{conv_to_scrm_arg.Feature_files}
\alias{conv_to_scrm_arg.Feature_growth}
\alias{conv_to_scrm_arg.Feature_migration}
\alias{conv_to_scrm_arg.Feature_mutation}
\alias{conv_to_scrm_arg.Feature_outgroup}
\alias{conv_to_scrm_arg.Feature_pop_merge}
\alias{conv_to_scrm_arg.Feature_recombination}
\alias{conv_to_scrm_arg.Feature_sample}
\alias{conv_to_scrm_arg.Feature_seg_sites}
\alias{conv_to_scrm_arg.Feature_selection}
\alias{conv_to_scrm_arg.Feature_size_change}
\alias{conv_to_scrm_arg.Feature_trees}
\alias{conv_to_scrm_arg.Feature_unphased}
\alias{conv_to_scrm_arg.default}
\alias{conv_to_seqgen_arg.Feature_files}
\alias{conv_to_seqgen_arg.Feature_growth}
\alias{conv_to_seqgen_arg.Feature_migration}
\alias{conv_to_seqgen_arg.Feature_mutation}
\alias{conv_to_seqgen_arg.Feature_outgroup}
\alias{conv_to_seqgen_arg.Feature_pop_merge}
\alias{conv_to_seqgen_arg.Feature_recombination}
\alias{conv_to_seqgen_arg.Feature_sample}
\alias{conv_to_seqgen_arg.Feature_seg_sites}
\alias{conv_to_seqgen_arg.Feature_selection}
\alias{conv_to_seqgen_arg.Feature_size_change}
\alias{conv_to_seqgen_arg.Feature_trees}
\alias{conv_to_seqgen_arg.Feature_unphased}
\alias{conv_to_seqgen_arg.default}
\title{Generate command line arguments for features}
\usage{
\method{conv_to_ms_arg}{Feature_growth}(feature, model)

\method{conv_to_msms_arg}{Feature_growth}(feature, model)

\method{conv_to_scrm_arg}{Feature_growth}(feature, model)

\method{conv_to_seqgen_arg}{Feature_growth}(feature, model)

\method{conv_to_ms_arg}{Feature_migration}(feature, model)

\method{conv_to_msms_arg}{Feature_migration}(feature, model)

\method{conv_to_scrm_arg}{Feature_migration}(feature, model)

\method{conv_to_seqgen_arg}{Feature_migration}(feature, model)

\method{conv_to_ms_arg}{Feature_mutation}(feature, model)

\method{conv_to_msms_arg}{Feature_mutation}(feature, model)

\method{conv_to_scrm_arg}{Feature_mutation}(feature, model)

\method{conv_to_seqgen_arg}{Feature_mutation}(feature, model)

\method{conv_to_ms_arg}{Feature_outgroup}(feature, model)

\method{conv_to_msms_arg}{Feature_outgroup}(feature, model)

\method{conv_to_scrm_arg}{Feature_outgroup}(feature, model)

\method{conv_to_seqgen_arg}{Feature_outgroup}(feature, model)

\method{conv_to_ms_arg}{Feature_pop_merge}(feature, model)

\method{conv_to_msms_arg}{Feature_pop_merge}(feature, model)

\method{conv_to_scrm_arg}{Feature_pop_merge}(feature, model)

\method{conv_to_seqgen_arg}{Feature_pop_merge}(feature, model)

\method{conv_to_ms_arg}{Feature_recombination}(feature, model)

\method{conv_to_msms_arg}{Feature_recombination}(feature, model)

\method{conv_to_scrm_arg}{Feature_recombination}(feature, model)

\method{conv_to_seqgen_arg}{Feature_recombination}(feature, model)

\method{conv_to_ms_arg}{Feature_sample}(feature, model)

\method{conv_to_msms_arg}{Feature_sample}(feature, model)

\method{conv_to_scrm_arg}{Feature_sample}(feature, model)

\method{conv_to_seqgen_arg}{Feature_sample}(feature, model)

\method{conv_to_ms_arg}{Feature_selection}(feature, model)

\method{conv_to_scrm_arg}{Feature_selection}(feature, model)

\method{conv_to_msms_arg}{Feature_selection}(feature, model)

\method{conv_to_seqgen_arg}{Feature_selection}(feature, model)

\method{conv_to_ms_arg}{Feature_size_change}(feature, model)

\method{conv_to_msms_arg}{Feature_size_change}(feature, model)

\method{conv_to_scrm_arg}{Feature_size_change}(feature, model)

\method{conv_to_seqgen_arg}{Feature_size_change}(feature, model)

\method{conv_to_ms_arg}{Feature_seg_sites}(feature, model)

\method{conv_to_msms_arg}{Feature_seg_sites}(feature, model)

\method{conv_to_scrm_arg}{Feature_seg_sites}(feature, model)

\method{conv_to_seqgen_arg}{Feature_seg_sites}(feature, model)

\method{conv_to_ms_arg}{Feature_trees}(feature, model)

\method{conv_to_msms_arg}{Feature_trees}(feature, model)

\method{conv_to_scrm_arg}{Feature_trees}(feature, model)

\method{conv_to_seqgen_arg}{Feature_trees}(feature, model)

\method{conv_to_ms_arg}{Feature_files}(feature, model)

\method{conv_to_msms_arg}{Feature_files}(feature, model)

\method{conv_to_scrm_arg}{Feature_files}(feature, model)

\method{conv_to_seqgen_arg}{Feature_files}(feature, model)

\method{conv_to_ms_arg}{Feature_unphased}(feature, model)

\method{conv_to_msms_arg}{Feature_unphased}(feature, model)

\method{conv_to_scrm_arg}{Feature_unphased}(feature, model)

\method{conv_to_seqgen_arg}{Feature_unphased}(feature, model)

conv_to_ms_arg(feature, model)

\method{conv_to_ms_arg}{default}(feature, model)

\method{conv_to_msms_arg}{default}(feature, model)

\method{conv_to_scrm_arg}{default}(feature, model)

\method{conv_to_seqgen_arg}{default}(feature, model)
}
\arguments{
\item{feature}{The feature for which the argument is generated}

\item{model}{The complete model for which the argument is generated}
}
\description{
These functions are exported only for technical reasons
(because they are S3 methods) and are not intended for
users.
}
\section{Methods (by class)}{
\itemize{
\item \code{Feature_growth}: Feature conversion

\item \code{Feature_growth}: Feature conversion

\item \code{Feature_growth}: Feature conversion

\item \code{Feature_growth}: Feature conversion

\item \code{Feature_migration}: Feature conversion

\item \code{Feature_migration}: Feature conversion

\item \code{Feature_migration}: Feature conversion

\item \code{Feature_migration}: Feature conversion

\item \code{Feature_mutation}: Feature conversion

\item \code{Feature_mutation}: Feature conversion

\item \code{Feature_mutation}: Feature conversion

\item \code{Feature_mutation}: Feature conversion

\item \code{Feature_outgroup}: Feature conversion

\item \code{Feature_outgroup}: Feature conversion

\item \code{Feature_outgroup}: Feature conversion

\item \code{Feature_outgroup}: Feature conversion

\item \code{Feature_pop_merge}: Feature conversion

\item \code{Feature_pop_merge}: Feature conversion

\item \code{Feature_pop_merge}: Feature conversion

\item \code{Feature_pop_merge}: Feature conversion

\item \code{Feature_recombination}: Feature conversion

\item \code{Feature_recombination}: Feature conversion

\item \code{Feature_recombination}: Feature conversion

\item \code{Feature_recombination}: Feature conversion

\item \code{Feature_sample}: Feature conversion

\item \code{Feature_sample}: Feature conversion

\item \code{Feature_sample}: Feature conversion

\item \code{Feature_sample}: Feature conversion

\item \code{Feature_selection}: Feature conversion

\item \code{Feature_selection}: Feature conversion

\item \code{Feature_selection}: Feature conversion

\item \code{Feature_selection}: Feature conversion

\item \code{Feature_size_change}: Feature conversion

\item \code{Feature_size_change}: Feature conversion

\item \code{Feature_size_change}: Feature conversion

\item \code{Feature_size_change}: Feature conversion

\item \code{Feature_seg_sites}: Feature conversion

\item \code{Feature_seg_sites}: Feature conversion

\item \code{Feature_seg_sites}: Feature conversion

\item \code{Feature_seg_sites}: Feature conversion

\item \code{Feature_trees}: Feature conversion

\item \code{Feature_trees}: Feature conversion

\item \code{Feature_trees}: Feature conversion

\item \code{Feature_trees}: Feature conversion

\item \code{Feature_files}: Feature conversion

\item \code{Feature_files}: Feature conversion

\item \code{Feature_files}: Feature conversion

\item \code{Feature_files}: Feature conversion

\item \code{Feature_unphased}: Feature conversion

\item \code{Feature_unphased}: Feature conversion

\item \code{Feature_unphased}: Feature conversion

\item \code{Feature_unphased}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion
}}

