% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsites.R
\name{create_locus_trio}
\alias{create_locus_trio}
\title{Combines three segregating sites to a locus trio}
\usage{
create_locus_trio(left, middle, right)
}
\arguments{
\item{left}{The segregating sites from the left locus}

\item{middle}{The segregating sites from the middle locus}

\item{right}{The segregating sites from the right locus}
}
\description{
Combines three segregating sites to a locus trio
}

