% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.addLocus}
\alias{dm.addLocus}
\title{Defines how many identical loci belong to a group of loci}
\usage{
dm.addLocus(dm, length, number = 1, group = 0)
}
\arguments{
\item{dm}{The Demographic Model}

\item{length}{The average length of the loci or the acctual of the locus
if just one is used.}

\item{number}{The number of loci to add to the group.}

\item{group}{The group for which we set the loci number}
}
\value{
The changed Demographic Model
}
\description{
Defines how many identical loci belong to a group of loci
}
\examples{
dm <- CoalModel(c(25,25), 100)
dm <- dm.addLocus(dm, number = 200, length = 250, group = 1)
}

