% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.addLocusTrio}
\alias{dm.addLocusTrio}
\title{Adds a trio of loci to a group}
\usage{
dm.addLocusTrio(dm, locus_names = c(left = "", middle = "", right = ""),
  locus_length = c(left = 1000, middle = 1000, right = 1000),
  distance = c(left_middle = 500, middle_right = 500), group = 1)
}
\arguments{
\item{dm}{The Demographic Model}

\item{locus_names}{A vector of 3 strings, giving the names for the loci.
The names are used for identifying the loci later (left, middle and right).}

\item{locus_length}{An integer vector of length 3, giving the length of each
of the three loci (left, middle and right).}

\item{distance}{A vector of two, giving the distance between left and middle,
and middle an right locus, in basepairs.}

\item{group}{The group to which to add the trio}
}
\value{
The extended demographic model
}
\description{
Adds a trio of loci to a group
}
\examples{
dm <- dm.createDemographicModel(c(25,25), 100)
dm <- dm.addLocusTrio(dm, locus_names = c('Solyc00g00500.2',
                                          'Solyc00g00520.1',
                                          'Solyc00g00540.1'),
                      locus_length=c(1250, 1017, 980),
                      distance=c(257, 814))
}

