% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.addPositiveSelection}
\alias{dm.addPositiveSelection}
\title{Adds positiv selection to a model}
\usage{
dm.addPositiveSelection(dm, min.strength = NA, max.strength = NA,
  parameter = "s", variance = 0, fraction.neutral = 0, population,
  at.time, group = 0)
}
\arguments{
\item{min.strength}{Minimal strength of selection}

\item{max.strength}{Maximal strength of selection}

\item{fraction.neutral}{Optionally, a fraction of the loci in the group can
be neutral.}

\item{population}{The populaton in which the allele is selected.}

\item{at.time}{The time at which the selection starts.}
}
\description{
Adds positiv selection to a model
}

