% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.getGroups}
\alias{dm.getGroups}
\title{Returns the groups currently in the model}
\usage{
dm.getGroups(dm)
}
\arguments{
\item{dm}{The demographic model}
}
\value{
The groups in the model.
}
\description{
Returns the groups currently in the model
}

