% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.getLociLength}
\alias{dm.getLociLength}
\title{Gets how long the loci in a group are}
\usage{
dm.getLociLength(dm, group = 1)
}
\arguments{
\item{dm}{The Demographic Model}

\item{group}{The group for which we get the length of loci}
}
\value{
The length of the loci in the group
}
\description{
Gets how long the loci in a group are
}
\examples{
dm <- dm.createDemographicModel(c(25,25), 100)
dm <- dm.addLocus(dm, number = 200, length = 250, group = 1)
dm.getLociLength(dm)
dm.getLociLength(dm, group = 0)
dm.getLociLength(dm, group = 2)
}

