% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.getLociNumber}
\alias{dm.getLociNumber}
\title{Gets how many loci belong to a group of loci}
\usage{
dm.getLociNumber(dm, group = 1)
}
\arguments{
\item{dm}{The Demographic Model}

\item{group}{The group for which we get the number of loci. Defaults to
the first group.}
}
\value{
The number of loci in the group
}
\description{
Gets how many loci belong to a group of loci
}
\examples{
dm <- dm.createDemographicModel(c(25,25), 100)
dm <- dm.addLocus(dm, number = 200, length = 250, group = 1)
dm.getLociNumber(dm)
dm.getLociNumber(dm, group = 0)
}

