% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{dm.setTrioMutationRates}
\alias{dm.setTrioMutationRates}
\title{Set the mutation rates for trios}
\usage{
dm.setTrioMutationRates(dm, middle_rate, outer_rate, group = 0)
}
\arguments{
\item{middle_rate}{The mutation rate used for the middle locus}

\item{outer_rate}{The mutation rate for the two outer loci}
}
\description{
Set the mutation rates for trios
}

