% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_migration.R
\name{feat_migration}
\alias{feat_migration}
\title{Feature: Migration/Gene Flow}
\usage{
feat_migration(rate, pop_from = NULL, pop_to = NULL, symmetric = FALSE,
  time = "0")
}
\arguments{
\item{rate}{The migration rate. Can be a numeric or a
\code{\link{parameter}}.}

\item{pop_from}{The population from which the individuals leave.}

\item{pop_to}{The population to which the individuals move.}

\item{symmetric}{Use the rate between all pairs of populations.}

\item{time}{The time point at which the migration with this rate starts.}
}
\description{
This feature adds migration between two populations to the model.
}
\details{
You can enter a time (\code{time}) at which the migration is
assumed to start (looking backwards in time). From that time on, a
fixed number of migrants move from population \code{pop_from} to
population \code{pop_to} each generation. This number is given via this
feature's parameter, which equals 4*N0*m,  where m is the
fraction of \code{pop_to}that is replaced with migrants each generation.
If \code{pop_to} has also size Ne, than this is just the
expected number of individuals that migrate each generation.

You can add different mutation rates at different times to your model.
Then each rate will be used for the period from its time point to
the next. Migration from and to an population always ends with the
speciation event in which the population is created.
}
\examples{
# Asymmetric migration between two populations:
model <- coal_model(c(5, 5), 10) +
  feat_migration(0.5, 1, 2) +
  feat_migration(1.0, 2, 1) +
  feat_mutation(5) +
  sumstat_sfs()
simulate(model)

# Three populations that exchange migrations with equal
# rates at times more than 0.5 time units in the past:
model <- coal_model(c(3, 4, 5), 2) +
  feat_migration(1.2, symmetric = TRUE, time = 0.5) +
  feat_mutation(5) +
  sumstat_sfs()
simulate(model)
}

