% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_outgroup.R
\name{feat_outgroup}
\alias{feat_outgroup}
\title{Adds an outgroup to a demographic model}
\usage{
feat_outgroup(population)
}
\arguments{
\item{population}{The population that is marked as outgroup.}
}
\description{
This feature declares an existing population as outgroup. Outgroups are used
to determine the ancestral allele in finite sites simulations.
}
\examples{
# A simple finite sites model
model <- coal_model(c(4, 6, 1), 2, 10) +
   feat_outgroup(3) +
   feat_pop_merge(par_range('tau', 0.5, 2), 2, 1) +
   feat_pop_merge(par_expr('2*tau'), 3, 1) +
   feat_mutation(par_range('theta', 1, 10), model="HKY")
}

