% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/feature_pop_merge.R
\name{feat_pop_merge}
\alias{feat_pop_merge}
\title{Adds a speciation event to a demographic model}
\usage{
feat_pop_merge(time, pop_source, pop_target)
}
\arguments{
\item{in.pop}{The population in which the spilt
occurs. See above for more information.}

\item{to.pop}{The newly created population.}

\item{time.point}{Instead of creating a new parameter, you can also
set the mutation rate to an expression based on existing
parameters. For example setting this to "tau" will use
an parameter with name tau that you have previously
created. You can also use R expression here, i.e. "2*tau"
or "5*M+2*tau" (if M is another parameter) will also
work (also this does not make much sense).}
}
\value{
The demographic model with a split.
}
\description{
You can use this function the create a new population that splits
of from an existing population at a given time in the past. The time
can be given as parameter or as an expression based on previously
generated time points.
}
\details{
Time in measured in Number of 4N0 generations in the past,
where N0 is the size of population 1 at time 0.
}
\examples{
dm <- dm.createDemographicModel(c(25,25), 100) +
  feat_pop_merge(par_range('tau', 0.01, 5), 2, 1) +
  feat_mutation(par_range('theta', 1, 20))
}

