% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_recombination.R
\name{feat_recombination}
\alias{feat_recombination}
\title{Feature: Recombination}
\usage{
feat_recombination(rate)
}
\arguments{
\item{rate}{The recombination rate. Can be a numeric or a
\code{\link{parameter}}. The rate is equal to
\eqn{4*N0*r}, where \eqn{r} is the probability that a
recombination event within the locus occurs in one generation.}
}
\value{
The feature, which can be added to a model using `+`.
}
\description{
This feature adds intra-locus recombination to a model.
}
\examples{
# Simulate a 1.5kb sequence for 10 individuals with recombination:
model <- coal_model(10, 2, 1500) +
  feat_recombination(1.5) +
  feat_mutation(5) +
  sumstat_sfs()
simulate(model)
}

