% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_sample.R
\name{feat_sample}
\alias{feat_sample}
\title{Creates a feature that represents the sampling from one population}
\usage{
feat_sample(size, population, time_point = "0")
}
\arguments{
\item{size}{The number of individuals that are sampled.}

\item{population}{The population from with the indidivuals are sampled}

\item{time_point}{The time at which the sample is taken.}
}
\value{
The feature, which can be added to a model using `+`.
}
\description{
Creates a feature that represents the sampling from one population
}

