% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feat_selection}
\alias{feat_selection}
\title{Adds positiv selection to a model}
\usage{
feat_selection(strength_AA, strength_Aa, variance = 0, fraction_neutral = 0,
  population, at_time, group = 0)
}
\arguments{
\item{fraction_neutral}{Optionally, a fraction of the loci in the group can
be neutral.}

\item{population}{The populaton in which the allele is selected.}

\item{at_time}{The time at which the selection starts.}
}
\description{
Adds positiv selection to a model
}

