% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_size_change.R
\name{feat_size_change}
\alias{feat_size_change}
\title{Adds an instantaneous change of the population size of one
population to a model.}
\usage{
feat_size_change(new_size, population, at.time = "0")
}
\arguments{
\item{new_size}{A \code{\link{Parameter}} giving the new size of the population
as fraction of N0.}

\item{population}{The number of the population whichs size is changed.}

\item{at.time}{The time point at which the size changes.}

\item{model}{The demographic model to which the size change should be added.}
}
\value{
A feature which can be added to a model.
}
\description{
This function changes the effective population size of one
population. The change is performed at a given time point
('at.time') and applies to the time interval farther into
the past from this point. The population size is set to a
fraction of N0, the present day size of population one.
}
\details{
If you want to add a slow, continuous change over some time,
then use the \link{model.addGrowth} function.
}
\examples{
# A model with one smaller population
model <- coal_model(c(20,37), 88) +
  feat_size_change(par_const(.1), 2, at.time='1')
}

