% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_unphased.R
\name{feat_unphased}
\alias{feat_unphased}
\title{Creates a feature that represents the sampling from one population}
\usage{
feat_unphased(ploidy, samples_per_ind = ploidy)
}
\arguments{
\item{size}{The number of individuals that are sampled.}

\item{population}{The population from with the indidivuals are sampled}

\item{time_point}{The time at which the sample is taken.}
}
\value{
The feature, which can be added to a model using `+`.
}
\description{
Creates a feature that represents the sampling from one population
}

