% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_getters.R
\name{get_feature_table}
\alias{get_feature_table}
\alias{get_groups}
\alias{get_locus_length}
\alias{get_locus_length_matrix}
\alias{get_locus_number}
\alias{get_outgroup}
\alias{get_outgroup_size}
\alias{get_parameter_table}
\alias{get_populations}
\alias{get_sample_size}
\alias{get_summary_statistics}
\title{Getters for Coalescent models}
\usage{
get_feature_table(model)

get_parameter_table(model)

get_locus_length(dm, group = 1)

get_locus_number(dm, group = 1)

get_populations(dm)

get_sample_size(dm)

get_summary_statistics(dm, group = 1, pop)

get_groups(dm)

get_locus_length_matrix(dm, group = 1)

get_outgroup(model)

get_outgroup_size(model)
}
\arguments{
\item{model}{The coalescent model from which aspects are returned}

\item{group}{The locus group for which aspects are returned}

\item{pop}{The population for which aspects are returned}
}
\description{
Getters for Coalescent models
}
\section{Functions}{
\itemize{
\item \code{get_feature_table}: Returns the features of a model as a data.frame

\item \code{get_parameter_table}: Returns the ranged parameters of a model as a
data.frame

\item \code{get_locus_length}: Returns the length of the loci in a locus group

\item \code{get_locus_number}: Returns the number of loci in a locus group

\item \code{get_populations}: Returns a vector of populations in the model

\item \code{get_sample_size}: Returns a vector of samples sizes per population.

\item \code{get_summary_statistics}: Returns a vector of populations in the model

\item \code{get_groups}: Returns a vector of groups in the model

\item \code{get_locus_length_matrix}: Returns a matrix with detailed length
information about the loci in the model.

\item \code{get_outgroup}: Returns the population that is marked as outgroup

\item \code{get_outgroup_size}: Returns the number of samples in the outgroup
}}
\author{
Paul Staab
}

