% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_program_seqgen.R
\name{Jaatha.setSeqgenExecutable}
\alias{Jaatha.setSeqgenExecutable}
\title{Set the path to the executable for seqgen}
\usage{
Jaatha.setSeqgenExecutable(seqgen.exe)
}
\arguments{
\item{seqgen.exe}{Path to seqgen's executable.}
}
\description{
Set the path to the executable for seqgen
}

