% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_class.R
\name{list_simulators}
\alias{list_simulators}
\title{Returns the available simulators}
\usage{
list_simulators()
}
\description{
This functions returns the usable simulators
}
\examples{
list_simulators()
}
