% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locus.R
\name{locus_trio}
\alias{locus_trio}
\title{Adds a trio of loci to a group}
\usage{
locus_trio(locus_length = c(left = 1000, middle = 1000, right = 1000),
  distance = c(left_middle = 500, middle_right = 500), number = 1)
}
\arguments{
\item{locus_length}{An integer vector of length 3, giving the length of each
of the three loci (left, middle and right).}

\item{distance}{A vector of two, giving the distance between left and middle,
and middle an right locus, in base pairs.}

\item{number}{The number of loci to add.}
}
\description{
Adds a trio of loci to a group
}
\examples{
coal_model(c(25,25)) +
  locus_trio(locus_length=c(1250, 1017, 980),
             distance=c(257, 814))
}

