% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_variation.R
\name{par_variation}
\alias{par_variation}
\title{Let the parameter values vary between different loci}
\usage{
par_variation(par, variance)
}
\arguments{
\item{par}{A parameter whose value will be set variable.}

\item{variance}{The variance of the gamma distribution, which the values used
for simulation will follow.}
}
\description{
The function can be used to let the values of a parameter vary between
the different loci. When used, the values for the enclosed parameter
will follow a gamma distribution with mean of the parameters original
value, and the variance specified as argument \code{variance}. This requires
that the original value is always positive. When using this, the simulators
are called separately for each locus, which can dramatically increase the
time needed to simulate models with many loci.
}

