% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_build.R
\name{+.coalmodelpart}
\alias{+.coalmodelpart}
\title{Add a feature or parameter to a model}
\usage{
\method{+}{coalmodelpart}(e1, e2)
}
\arguments{
\item{e1}{The Model to which the feature/parameter should be added}

\item{e2}{The feature/parameter to add}
}
\value{
The extended model
}
\description{
Add a feature or parameter to a model
}
\keyword{internal}
