% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_build.R
\name{+.CSR_OBJ}
\alias{+.CSR_OBJ}
\title{Add a feature or parameter to a model}
\usage{
\method{+}{CSR_OBJ}(e1, e2)
}
\arguments{
\item{e1}{The Model to which the feature/parameter should be added}

\item{e2}{The feature/parameter to add}
}
\value{
The extended model
}
\description{
Add a feature or parameter to a model
}

