% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_executable.R
\name{search_executable}
\alias{search_executable}
\title{Search the working directory and the run path for an executable}
\usage{
search_executable(name, envir_var = NULL)
}
\arguments{
\item{name}{The name of the executable to look for}

\item{envir_var}{the name of the environment variable to use}
}
\value{
The complete path of the executable is found, or "NULL" if not.
}
\description{
This function tries to find a binary of a given name by looking
in the current working directory and the directories listed in
the $PATH environment variable. If an environment variable with
name equal to to binaries name in upper case is given, it also
tries to use this as path of the binary.
}
\keyword{internal}
