% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_simulate.R
\name{simulate.Coalmodel}
\alias{simulate.Coalmodel}
\title{Simulates data according to a demographic model}
\usage{
\method{simulate}{Coalmodel}(object, nsim = 1, seed, pars = numeric(0), ...)
}
\arguments{
\item{model}{The demographic model according to which the simulations should be done}

\item{parameters}{A vector of parameters which should be used for the simulations.
If a matrix is given, a simulation for each row of the matrix will be performed}
}
\value{
A matrix where each row is the vector of summary statistics for
        the parameters in the same row of the "parameter" matrix
}
\description{
Simulates data according to a demographic model
}
\examples{
model <- coal_model(c(5,10), 20) +
  feat_pop_merge(par_range('tau', 0.01, 5), 2, 1) +
  feat_mutation(par_range('theta', 1, 10)) +
  sumstat_jsfs()

simulate(model, pars=c(1, 5))
}

