% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat.R
\docType{data}
\name{sumstat_class}
\alias{sumstat_class}
\title{Base Class for Summary Statistics}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x38537f0>
 - attr(*, "name")= chr "sumstat_generator"
}}
\usage{
sumstat_class
}
\description{
If you want to create additional summary statistics for coala, create
\pkg{R6} classes that inherit from this object.
}
\keyword{datasets}

