% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_file.R
\name{sumstat_file}
\alias{sumstat_file}
\title{Returns files with the raw results of simulations}
\usage{
sumstat_file(folder)
}
\arguments{
\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
Returns files with the raw results of simulations
}

