% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_ihh.R
\name{sumstat_ihh}
\alias{sumstat_ihh}
\title{Integrated Extended Haplotype Homozygosity}
\usage{
sumstat_ihh(name = "ihh", position = NA, population = 1, group = 0)
}
\arguments{
\item{position}{A position relative to the locus extent, e.g. 0.5 for the
middle of the locus. If provided, the iHH will be calculate using the SNP
closest to the given position as focal SNP. Otherwise, all SNPs will be
used as focal SNPs in turn, and all the values reportet. If trios are used,
than the position is relative to the middle locus' extend.}

\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
This summary statistic calculates a modified version of the iHH statistic
introduced by
}
\details{
Voight et al., A map of recent positive selection in the human genome.
 PLoS Biol, 4(3):e72, Mar 2006

coalsimr relies on the package rehh to calculate this statistic. Please refer
to their documentation for detailed information on the concrete
implementation.
}

