% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_omega.R
\name{sumstat_omega}
\alias{sumstat_omega}
\title{Calculates the Omega Statistic from Simulations}
\usage{
sumstat_omega(name = "omega", min_win = 100, max_win = 1000,
  grid = 1000, binary = "automatic")
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{min_win}{The minimum distance from the grid point that a SNP must have
to be included in the calculation of omega.}

\item{max_win}{The maximum distance from the grid point that a SNP must have
to be included in the calculation of omega.}

\item{grid}{The number of points for which omega is calculated on each
locus. Should be significantly lower than the locus length.}

\item{binary}{The path of the binary for OmegaPlus. If set to "automatic",
coala will try to find a binary called "OmegaPlus" using the PATH
environment variable.}
}
\description{
Calculates the Omega Statistic from Simulations
}

