% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_omegaplus.R
\name{sumstat_omegaplus}
\alias{sumstat_omegaplus}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_omegaplus(name = "omegaplus", min_win = 100, max_win = 1000,
  grid = 10000, binary = "automatic")
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}
}
\description{
Returns the Segregation Sites Statistics from simulations
}

