% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_omegaprime.R
\name{sumstat_omegaprime}
\alias{sumstat_omegaprime}
\title{Calculates the (experimental) Omega' Statistic}
\usage{
sumstat_omegaprime(name = "omegaprime", population = 1)
}
\description{
Calculates the (experimental) Omega' Statistic
}

