% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_seg_sites.R
\name{sumstat_seg_sites}
\alias{sumstat_seg_sites}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_seg_sites(name = "seg_sites", group = 0)
}
\arguments{
\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
Returns the Segregation Sites Statistics from simulations
}

