% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_sfs.R
\name{sumstat_sfs}
\alias{sumstat_sfs}
\title{Calculates the Site Frequency Spectrum from simulations}
\usage{
sumstat_sfs(name = "sfs", population = "all", group = 0)
}
\arguments{
\item{population}{Either the number of a population for with the SFS is
calculated, or \code{'all'} to calculate the combined SFS of all
populations.}

\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
Calculates the Site Frequency Spectrum from simulations
}

