% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_tajimas_d.R
\name{sumstat_tajimas_d}
\alias{sumstat_tajimas_d}
\title{Tajima's D}
\usage{
sumstat_tajimas_d(name = "tajimas_d", population = "all")
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{population}{The population for which the statistic is calculated.}
}
\value{
On simulation, this returns a vector with the value of Tajima's D for
  each locus.
}
\description{
This calculates Tajima's D from the simulation results when added to a model.
}
\details{
Tajima's D was introduced by

Tajima, F. (1989). "Statistical method for testing the neutral mutation
hypothesis by DNA polymorphism.". Genetics 123 (3): 585-95.
}

