% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_trees.R
\name{sumstat_trees}
\alias{sumstat_trees}
\title{Returns ancestral tress in NEWICK fromat from simulations}
\usage{
sumstat_trees(name = "trees")
}
\description{
Returns ancestral tress in NEWICK fromat from simulations
}

