% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_jsfs}
\alias{calc_jsfs}
\title{Calculates the JSFS for two populations}
\usage{
calc_jsfs(seg_sites, pop1, pop2)
}
\arguments{
\item{seg_sites}{List of segregating sites}

\item{pop1}{The rows of \code{seg_sites} that correspond to individuals
of the first population.}

\item{pop2}{same as \code{pop1}, but for the second population.}
}
\value{
The Joint Site Frequency Spectrum, as a matrix.
}
\description{
Calculates the JSFS for two populations
}

