% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_jsfs}
\alias{calc_jsfs}
\title{Calculates the Joint Site Frequency Spectrum}
\usage{
calc_jsfs(segsites_list, ind_per_pop)
}
\arguments{
\item{segsites_list}{List of segregating sites}

\item{ind_per_pop}{A list of integer vector, where each entry gives the
index of the haploids that belong the corresponding population.}
}
\value{
The Joint Site Frequency Spectrum, as a matrix.
}
\description{
Calculates the Joint Site Frequency Spectrum
}
\author{
Paul Staab & Dirk Metzler
}
\keyword{internal}
