% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{check_model}
\alias{check_model}
\title{Check which simulator can simulate a model}
\usage{
check_model(model)
}
\arguments{
\item{model}{The model which is checked}
}
\description{
This function checks which of the available simulators can
simulate a given model. It also states the problems for the ones that
are incompatible with the model.
}
\examples{
model <- coal_model(10, 1) +
  feat_mutation(5, fixed = TRUE)
check_model(model)
}
\seealso{
Do view the priority of the simulators: \code{\link{list_simulators}}
}
