% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coala.R
\docType{package}
\name{coala-package}
\alias{coala}
\alias{coala-package}
\title{A Framework for Coalescent Simulation in R}
\description{
This package allows to specify and simulate coalescent models from
within R. The \code{introduction} vignette is a good place to start.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statgenlmu/coala}
  \item Report bugs at \url{https://github.com/statgenlmu/coala/issues}
}

}
\author{
\strong{Maintainer}: Dirk Metzler \email{metzler@bio.lmu.de} [thesis advisor]

Authors:
\itemize{
  \item Paul Staab \email{develop@paulstaab.de}
}

Other contributors:
\itemize{
  \item Jorge E. Amaya Romero [contributor]
}

}
