% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coala.R
\docType{package}
\name{coala}
\alias{coala}
\alias{coala-package}
\title{A Framework for Coalescent Simulation in R}
\description{
This package allows to specify and simulate coalescent models from
within R. The `introduction` vignettes is a good place to start.
}
\author{
Paul R. Staab \email{develop (at) paulstaab.de}
}

