% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_growth.R,
%   R/feature_ignore_singletons.R, R/feature_migration.R, R/feature_mutation.R,
%   R/feature_outgroup.R, R/feature_pop_merge.R, R/feature_recombination.R,
%   R/feature_sample.R, R/feature_selection.R, R/feature_size_change.R,
%   R/feature_sumstats.R, R/feature_unphased.R, R/simulator_ms.R,
%   R/simulator_msms.R, R/simulator_scrm.R, R/simulator_seqgen.R
\name{conv_to_ms_arg.growth}
\alias{conv_to_ms_arg.growth}
\alias{conv_to_msms_arg.growth}
\alias{conv_to_scrm_arg.growth}
\alias{conv_to_seqgen_arg.growth}
\alias{conv_to_ms_arg.ign_singletons}
\alias{conv_to_msms_arg.ign_singletons}
\alias{conv_to_scrm_arg.ign_singletons}
\alias{conv_to_seqgen_arg.ign_singletons}
\alias{conv_to_ms_arg.migration}
\alias{conv_to_msms_arg.migration}
\alias{conv_to_scrm_arg.migration}
\alias{conv_to_seqgen_arg.migration}
\alias{conv_to_ms_arg.mutation}
\alias{conv_to_msms_arg.mutation}
\alias{conv_to_scrm_arg.mutation}
\alias{conv_to_seqgen_arg.mutation}
\alias{conv_to_ms_arg.outgroup}
\alias{conv_to_msms_arg.outgroup}
\alias{conv_to_scrm_arg.outgroup}
\alias{conv_to_seqgen_arg.outgroup}
\alias{conv_to_ms_arg.pop_merge}
\alias{conv_to_msms_arg.pop_merge}
\alias{conv_to_scrm_arg.pop_merge}
\alias{conv_to_seqgen_arg.pop_merge}
\alias{conv_to_ms_arg.recombination}
\alias{conv_to_msms_arg.recombination}
\alias{conv_to_scrm_arg.recombination}
\alias{conv_to_seqgen_arg.recombination}
\alias{conv_to_ms_arg.sample}
\alias{conv_to_msms_arg.sample}
\alias{conv_to_scrm_arg.sample}
\alias{conv_to_seqgen_arg.sample}
\alias{conv_to_ms_arg.selection}
\alias{conv_to_scrm_arg.selection}
\alias{conv_to_msms_arg.selection}
\alias{conv_to_seqgen_arg.selection}
\alias{conv_to_ms_arg.size_change}
\alias{conv_to_msms_arg.size_change}
\alias{conv_to_scrm_arg.size_change}
\alias{conv_to_seqgen_arg.size_change}
\alias{conv_to_ms_arg.seg_sites_feat}
\alias{conv_to_msms_arg.seg_sites_feat}
\alias{conv_to_scrm_arg.seg_sites_feat}
\alias{conv_to_seqgen_arg.seg_sites_feat}
\alias{conv_to_ms_arg.trees_feat}
\alias{conv_to_msms_arg.trees_feat}
\alias{conv_to_scrm_arg.trees_feat}
\alias{conv_to_seqgen_arg.trees_feat}
\alias{conv_to_ms_arg.files_feat}
\alias{conv_to_msms_arg.files_feat}
\alias{conv_to_scrm_arg.files_feat}
\alias{conv_to_seqgen_arg.files_feat}
\alias{conv_to_ms_arg.unphased}
\alias{conv_to_msms_arg.unphased}
\alias{conv_to_scrm_arg.unphased}
\alias{conv_to_seqgen_arg.unphased}
\alias{conv_to_ms_arg}
\alias{conv_to_ms_arg.default}
\alias{conv_to_msms_arg.default}
\alias{conv_to_scrm_arg.default}
\alias{conv_to_seqgen_arg.default}
\title{Generate command line arguments for features}
\usage{
\method{conv_to_ms_arg}{growth}(feature, model)

\method{conv_to_msms_arg}{growth}(feature, model)

\method{conv_to_scrm_arg}{growth}(feature, model)

\method{conv_to_seqgen_arg}{growth}(feature, model)

\method{conv_to_ms_arg}{ign_singletons}(feature, model)

\method{conv_to_msms_arg}{ign_singletons}(feature, model)

\method{conv_to_scrm_arg}{ign_singletons}(feature, model)

\method{conv_to_seqgen_arg}{ign_singletons}(feature, model)

\method{conv_to_ms_arg}{migration}(feature, model)

\method{conv_to_msms_arg}{migration}(feature, model)

\method{conv_to_scrm_arg}{migration}(feature, model)

\method{conv_to_seqgen_arg}{migration}(feature, model)

\method{conv_to_ms_arg}{mutation}(feature, model)

\method{conv_to_msms_arg}{mutation}(feature, model)

\method{conv_to_scrm_arg}{mutation}(feature, model)

\method{conv_to_seqgen_arg}{mutation}(feature, model)

\method{conv_to_ms_arg}{outgroup}(feature, model)

\method{conv_to_msms_arg}{outgroup}(feature, model)

\method{conv_to_scrm_arg}{outgroup}(feature, model)

\method{conv_to_seqgen_arg}{outgroup}(feature, model)

\method{conv_to_ms_arg}{pop_merge}(feature, model)

\method{conv_to_msms_arg}{pop_merge}(feature, model)

\method{conv_to_scrm_arg}{pop_merge}(feature, model)

\method{conv_to_seqgen_arg}{pop_merge}(feature, model)

\method{conv_to_ms_arg}{recombination}(feature, model)

\method{conv_to_msms_arg}{recombination}(feature, model)

\method{conv_to_scrm_arg}{recombination}(feature, model)

\method{conv_to_seqgen_arg}{recombination}(feature, model)

\method{conv_to_ms_arg}{sample}(feature, model)

\method{conv_to_msms_arg}{sample}(feature, model)

\method{conv_to_scrm_arg}{sample}(feature, model)

\method{conv_to_seqgen_arg}{sample}(feature, model)

\method{conv_to_ms_arg}{selection}(feature, model)

\method{conv_to_scrm_arg}{selection}(feature, model)

\method{conv_to_msms_arg}{selection}(feature, model)

\method{conv_to_seqgen_arg}{selection}(feature, model)

\method{conv_to_ms_arg}{size_change}(feature, model)

\method{conv_to_msms_arg}{size_change}(feature, model)

\method{conv_to_scrm_arg}{size_change}(feature, model)

\method{conv_to_seqgen_arg}{size_change}(feature, model)

\method{conv_to_ms_arg}{seg_sites_feat}(feature, model)

\method{conv_to_msms_arg}{seg_sites_feat}(feature, model)

\method{conv_to_scrm_arg}{seg_sites_feat}(feature, model)

\method{conv_to_seqgen_arg}{seg_sites_feat}(feature, model)

\method{conv_to_ms_arg}{trees_feat}(feature, model)

\method{conv_to_msms_arg}{trees_feat}(feature, model)

\method{conv_to_scrm_arg}{trees_feat}(feature, model)

\method{conv_to_seqgen_arg}{trees_feat}(feature, model)

\method{conv_to_ms_arg}{files_feat}(feature, model)

\method{conv_to_msms_arg}{files_feat}(feature, model)

\method{conv_to_scrm_arg}{files_feat}(feature, model)

\method{conv_to_seqgen_arg}{files_feat}(feature, model)

\method{conv_to_ms_arg}{unphased}(feature, model)

\method{conv_to_msms_arg}{unphased}(feature, model)

\method{conv_to_scrm_arg}{unphased}(feature, model)

\method{conv_to_seqgen_arg}{unphased}(feature, model)

conv_to_ms_arg(feature, model)

\method{conv_to_ms_arg}{default}(feature, model)

\method{conv_to_msms_arg}{default}(feature, model)

\method{conv_to_scrm_arg}{default}(feature, model)

\method{conv_to_seqgen_arg}{default}(feature, model)
}
\arguments{
\item{feature}{The feature for which the argument is generated}

\item{model}{The complete model for which the argument is generated}
}
\description{
These functions are exported only for technical reasons
(because they are S3 methods) and are not intended for
users.
}
\section{Methods (by class)}{
\itemize{
\item \code{growth}: Feature conversion

\item \code{growth}: Feature conversion

\item \code{growth}: Feature conversion

\item \code{growth}: Feature conversion

\item \code{ign_singletons}: Feature conversion

\item \code{ign_singletons}: Feature conversion

\item \code{ign_singletons}: Feature conversion

\item \code{ign_singletons}: Feature conversion

\item \code{migration}: Feature conversion

\item \code{migration}: Feature conversion

\item \code{migration}: Feature conversion

\item \code{migration}: Feature conversion

\item \code{mutation}: Feature conversion

\item \code{mutation}: Feature conversion

\item \code{mutation}: Feature conversion

\item \code{mutation}: Feature conversion

\item \code{outgroup}: Feature conversion

\item \code{outgroup}: Feature conversion

\item \code{outgroup}: Feature conversion

\item \code{outgroup}: Feature conversion

\item \code{pop_merge}: Feature conversion

\item \code{pop_merge}: Feature conversion

\item \code{pop_merge}: Feature conversion

\item \code{pop_merge}: Feature conversion

\item \code{recombination}: Feature conversion

\item \code{recombination}: Feature conversion

\item \code{recombination}: Feature conversion

\item \code{recombination}: Feature conversion

\item \code{sample}: Feature conversion

\item \code{sample}: Feature conversion

\item \code{sample}: Feature conversion

\item \code{sample}: Feature conversion

\item \code{selection}: Feature conversion

\item \code{selection}: Feature conversion

\item \code{selection}: Feature conversion

\item \code{selection}: Feature conversion

\item \code{size_change}: Feature conversion

\item \code{size_change}: Feature conversion

\item \code{size_change}: Feature conversion

\item \code{size_change}: Feature conversion

\item \code{seg_sites_feat}: Feature conversion

\item \code{seg_sites_feat}: Feature conversion

\item \code{seg_sites_feat}: Feature conversion

\item \code{seg_sites_feat}: Feature conversion

\item \code{trees_feat}: Feature conversion

\item \code{trees_feat}: Feature conversion

\item \code{trees_feat}: Feature conversion

\item \code{trees_feat}: Feature conversion

\item \code{files_feat}: Feature conversion

\item \code{files_feat}: Feature conversion

\item \code{files_feat}: Feature conversion

\item \code{files_feat}: Feature conversion

\item \code{unphased}: Feature conversion

\item \code{unphased}: Feature conversion

\item \code{unphased}: Feature conversion

\item \code{unphased}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion

\item \code{default}: Feature conversion
}}

\keyword{internal}
