% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_mutation.R
\name{feat_mutation}
\alias{feat_mutation}
\title{Feature: Mutation}
\usage{
feat_mutation(rate, model = "IFS", base_frequencies = NA, tstv_ratio = NA,
  gtr_rates = NA, fixed_number = FALSE)
}
\arguments{
\item{rate}{The mutation rate. Can be a numeric or a \code{\link{parameter}}.
The rate is specified as \eqn{4 * N0 * mu}, where \eqn{mu} is the
mutation rate per locus.}

\item{model}{The mutation model you want to use.
Can be either 'IFS' (default), 'HKY' or 'GTR'. Refer to the mutation model
section for detailed information.}

\item{base_frequencies}{The equilibrium frequencies of the four bases used in
the 'HKY' mutation model. Must be a numeric vector of length four, with the
values for A, C, G and T, in that order.}

\item{tstv_ratio}{The ratio of transitions to transversions used in the 'HKY'
muation model.}

\item{gtr_rates}{The rates for the six amino acid substitutions used in the
'GTR' model. Must be a numeric vector of length six.
Order: A<->C, A<->G, A<->T, C<->G, C<->T, G<->T.}

\item{fixed_number}{If set to \code{TRUE}, the number of mutations on each
locus will always be exactly equal to the rate, rather than happening with
a rate along the ancestral tree.}
}
\value{
The feature, which can be added to a model using `+`.
}
\description{
This feature adds mutations to a model. The mutations occur in the genomes
of the individuals with a given \code{rate}.
}
\section{Mutation Models}{

The infinite sites mutation (\strong{IFS}) model is a frequently used simplification
in population genetics. It assumes that each locus consists of infinitely
many sites at which mutations can occure, and each mutation hits a new site.
Consequently, there are no back-mutations with this model. It does not
generate DNA sequences, but rather only 0/1 coded data, were 0 denotes the
ancestral state of the site, and 1 the derived state created by a mutation.

The other mutation models are finite site models that generate more realistic
sequences.

The Hasegawa, Kishino and Yano (\strong{HKY}) model (Hasegawa et al., 1985) allows
for a different rate of transitions and transversions (tstv_ratio)
and unequal
frequencies of the four nucleotides (base_frequencies).

The general reversible process (\strong{GTR}) model (e.g. Yang, 1994) is more general
than the HKY model and allows to define the rates for each
type of substitution. The rates are assumed to be symmetric
(e.g., the rate for T to G is equal to the one for G to T).
}
\examples{
# A model with a constant mutation rate of 5:
model <- coal_model(10, 1) + feat_mutation(rate = 5)

# A model with a mutation rate that can be estimated with Jaatha:
model <- coal_model(c(15,20), 100) +
  feat_mutation(par_range('theta', 1, 20))
}

