% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_recombination.R
\name{feat_recombination}
\alias{feat_recombination}
\title{Feature: Recombination}
\usage{
feat_recombination(rate, locus_group = "all")
}
\arguments{
\item{rate}{The recombination rate. Can be a numeric or a
\code{\link{parameter}}. The rate is equal to
\eqn{4*N0*r}, where \eqn{r} is the probability that a
recombination event within the locus occurs in one generation.}

\item{locus_group}{The loci for which this features is used. Can either be
\code{"all"} (default), in which case the feature is used for simulating
all loci, or a numeric vector. In the latter case, the feature is only
used for the loci added in \code{locus_} commands  with the corresponding
index starting from 1 in order in which the commands where added to the
model. For example, if a model has
\code{locus_single(10) + locus_averaged(10, 11) + locus_single(12)} and
this argument is \code{c(2, 3)}, than the feature is used for all but
the first locus (that is locus 2 - 12).}
}
\value{
The feature, which can be added to a model using `+`.

The feature, which can be added to a model created with
  \code{\link{coal_model}} using \code{+}.
}
\description{
This feature adds intra-locus recombination to a model.  The rate is per locus
for \link[=locus]{unlinked loci} and per trio for linked
\link[=locus_trio]{locus trios}. By default, the same recombination rate is used
for all loci, but it is possible to change this with \code{\link{par_variation}}
and \code{\link{par_zero_inflation}}. Coala assumes that recombination events
can occur between all neighboring bases.
}
\examples{
# Simulate a 1.5kb sequence for 10 individuals with recombination:
model <- coal_model(10, 2, 1500) +
  feat_recombination(1.5) +
  feat_mutation(5) +
  sumstat_sfs()
simulate(model)
}
\seealso{
For creating a model: \code{\link{coal_model}}

For adding recombination: \code{\link{feat_mutation}}.

Other features: 
\code{\link{feat_growth}()},
\code{\link{feat_ignore_singletons}()},
\code{\link{feat_migration}()},
\code{\link{feat_mutation}()},
\code{\link{feat_outgroup}()},
\code{\link{feat_pop_merge}()},
\code{\link{feat_selection}()},
\code{\link{feat_size_change}()},
\code{\link{feat_unphased}()}
}
\concept{features}
