% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_sample.R
\name{feat_sample}
\alias{feat_sample}
\title{Creates a feature that represents the sampling from one population}
\usage{
feat_sample(sizes, time = "0")
}
\arguments{
\item{sizes}{The number of individuals that are sampled per population,
given as a numeric vector.}

\item{time}{The time at which the sample is taken.}
}
\value{
The feature, which can be added to a model using `+`.
}
\description{
Creates a feature that represents the sampling from one population
}

