% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sample.R
\name{feat_sample}
\alias{feat_sample}
\title{Creates a feature that represents the sampling from one population}
\usage{
feat_sample(individuals, ploidy = 1, time = "0")
}
\arguments{
\item{individuals}{The number of individuals that are sampled per population,
given as a numeric vector.}

\item{ploidy}{The number of chromosomes that will be simulated per
individual.}

\item{time}{The time at which the sample is taken.}
}
\description{
Creates a feature that represents the sampling from one population
}
\keyword{interna}
