% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_expr}
\alias{ModelParameters}
\alias{par_const}
\alias{par_expr}
\alias{par_range}
\title{Define Model Parameters}
\usage{
par_expr(expr)

par_const(constant)

par_range(name, lower, upper)
}
\arguments{
\item{expr}{An R command. The command will not be evaluted until a simulation
is performed. I can contain other named parameters, but not parameters
created with \code{par_expr}. Make sure that the expression always evaluates
to a suitable parameter value (a single numeric in almost all cases).}

\item{constant}{The constant value of the parameter}

\item{name}{A string. The name of the parameter.}

\item{lower}{A numeric. The lower boundary of the parameter's range.}

\item{upper}{A numeric. The upper boundary of the parameter's range.}
}
\description{
bla bla bla
}
\section{Functions}{
\itemize{
\item \code{par_expr}: A parameter whichs value is determined by evalutating an
expression.

\item \code{par_const}: Creates an parameter that is equal to a fixed value.
Different to par_expr, the value is evaluated on parameter creation.

\item \code{par_range}: Creates an parameter that can take a range of possible
values. Used for creating model parameters for \code{Jaatha}.
}}
\examples{
par_const(5)         # The parameters value is always 5.
par_expr(runif(1))   # Creates an parameter which takes a uniformly
                     # distributed value in each simulation.
par_range('x', 1, 5) # Creates an parameter with name x with possible values
                     # between 1 and 5.
par_expr(2*x)        # The parameters value is always equal two twice the
                     # value of a different model parameter named 'x'.
}
\author{
Paul Staab
}

