% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator_seqgen.R
\name{set_seqgen_exe}
\alias{set_seqgen_exe}
\title{Set the path to the executable for seqgen}
\usage{
set_seqgen_exe(seqgen_exe)
}
\arguments{
\item{seqgen_exe}{The path to seqgen's executable.}
}
\description{
Set the path to the executable for seqgen
}

