% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{simulate.CoalModel}
\alias{simulate.CoalModel}
\title{Simulates data according to a demographic model}
\usage{
\method{simulate}{CoalModel}(dm, parameters)
}
\arguments{
\item{dm}{The demographic model according to which the simulations should be done}

\item{parameters}{A vector of parameters which should be used for the simulations.
If a matrix is given, a simulation for each row of the matrix will be performed}
}
\value{
A matrix where each row is the vector of summary statistics for
        the parameters in the same row of the "parameter" matrix
}
\description{
Simulates data according to a demographic model
}
\examples{
model <- CoalModel(c(5,10), 20) +
  feat_pop_merge(par_range('tau', 0.01, 5), 2, 1) +
  feat_mutation(par_range('theta', 1, 10))
model <- dm.addSummaryStatistic(model, 'jsfs')

simulate(model, c(1, 5))
}

