% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_simulate.R
\name{simulate.coalmodel}
\alias{simulate.coalmodel}
\title{Simulates data according to a demographic model}
\usage{
\method{simulate}{coalmodel}(object, nsim = 1, seed, ..., pars = numeric(0))
}
\arguments{
\item{object}{The coalescent model to be simulated}

\item{nsim}{currently unused}

\item{seed}{A random seed that is set before simulation.}

\item{...}{currently unused}

\item{pars}{Values for parameters specified with \code{\link{par_named}} or
\code{\link{par_range}}. Should be a named numeric vector.}
}
\value{
A list of summary statistics.
}
\description{
Simulates data according to a demographic model
}
\examples{
model <- coal_model(c(5,10), 20) +
  feat_pop_merge(par_range('tau', 0.01, 5), 2, 1) +
  feat_mutation(par_range('theta', 1, 10)) +
  sumstat_jsfs()

simulate(model, pars=c(1, 5))
}

