% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_ms.R
\name{simulator_ms}
\alias{simulator_ms}
\alias{activate_ms}
\title{Simulator: ms}
\usage{
activate_ms(priority = 300)
}
\arguments{
\item{priority}{The priority for this simulator. If multiple simulators
can simulate a model, the one with the highest priority will be used.}
}
\description{
This function adds the simulator 'ms' to the list of available simulators.
In order to use 'ms', you need to install the CRAN package \pkg{phyclust}.
By default, 'scrm' will be preferred over 'ms'. Raise the priority of 'ms'
to change this behavior.
}
\examples{
# To prefer ms to scrm:
\dontrun{activate_ms(priority = 500)}
}
\references{
Richard R. Hudson.
Generating samples under a Wright-Fisher neutral model of genetic variation.
Bioinformatics (2002) 18 (2): 337-338
doi:10.1093/bioinformatics/18.2.337
}
\seealso{
Other simulators: 
\code{\link{simulator_msms}},
\code{\link{simulator_scrm}},
\code{\link{simulator_seqgen}}
}
\concept{simulators}
