% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat.R
\docType{data}
\name{sumstat_class}
\alias{sumstat_class}
\title{Base Class for Summary Statistics}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sumstat_class
}
\description{
If you want to create additional summary statistics for coala, create
\pkg{R6} classes that inherit from this object.
}
\keyword{internal}

