% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_dna.R
\name{sumstat_dna}
\alias{sumstat_dna}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_dna(name = "dna")
}
\description{
Returns the Segregation Sites Statistics from simulations
}

