% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_dna.R
\name{sumstat_dna}
\alias{sumstat_dna}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_dna(name = "dna")
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}
}
\description{
Returns the Segregation Sites Statistics from simulations
}

