% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_dna.R
\name{sumstat_dna}
\alias{sumstat_dna}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_dna(name = "dna", transformation = identity)
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{transformation}{An optional function for transforming the results
of the statistic. If specified, the results of the transformation are
returned instead of the original values.}
}
\description{
Returns the Segregation Sites Statistics from simulations
}

