% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_file.R
\name{sumstat_file}
\alias{sumstat_file}
\title{Returns files with the raw results of simulations}
\usage{
sumstat_file(folder)
}
\arguments{
\item{folder}{The path of a folder where the files will be written.}
}
\description{
Returns files with the raw results of simulations
}

