% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_ihh.R
\name{sumstat_iHH}
\alias{sumstat_iHH}
\title{Integrated Extended Haplotype Homozygosity}
\usage{
sumstat_iHH(name = "iHH", group = 0)
}
\arguments{
\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
This summary statistic calculates a modified version of the iHH statistic
introduced by
}
\details{
Voight et al., A map of recent positive selection in the human genome.
 PLoS Biol, 4(3):e72, Mar 2006

their documentation for detailed information.
}

