% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_jsfs.R
\name{sumstat_jsfs}
\alias{sumstat_jsfs}
\title{Returns the Joint Site Frequency Spectrum from simulations}
\usage{
sumstat_jsfs(group = 0)
}
\arguments{
\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
Returns the Joint Site Frequency Spectrum from simulations
}

