% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_jsfs.R
\name{sumstat_jsfs}
\alias{sumstat_jsfs}
\title{Calculates the Joint Site Frequency Spectrum from simulations}
\usage{
sumstat_jsfs(name = "jsfs", populations = c(1, 2))
}
\description{
Calculates the Joint Site Frequency Spectrum from simulations
}

