% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_jsfs.R
\name{sumstat_jsfs}
\alias{sumstat_jsfs}
\title{Calculates the Joint Site Frequency Spectrum from simulations}
\usage{
sumstat_jsfs(name = "jsfs", populations = c(1, 2))
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{populations}{The populations for which the statistic is calculated.}
}
\description{
Calculates the Joint Site Frequency Spectrum from simulations
}

