% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_mcmf.R
\name{sumstat_mcmf}
\alias{sumstat_mcmf}
\title{The MCMF Summary Statistic}
\usage{
sumstat_mcmf(name = "mcmf", population = 1)
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{population}{The population for which the statistic is calculated.}
}
\description{
When adding this to a model, the "Most Common Mutation Frequecy" summary
statistic is calculated from the simulation output.
}

