% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumstat_omegaprime.R
\name{sumstat_omegaprime}
\alias{sumstat_omegaprime}
\title{Calculates the (experimental) Omega" Statistic}
\usage{
sumstat_omegaprime(name = "omegaprime", population = 1)
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{population}{The population for which the statistic is calculated.}
}
\description{
Calculates the (experimental) Omega" Statistic
}

