% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_seg_sites.R
\name{sumstat_seg_sites}
\alias{sumstat_seg_sites}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_seg_sites(name = "seg_sites")
}
\description{
Returns the Segregation Sites Statistics from simulations
}

