% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_seg_sites.R
\name{sumstat_seg_sites}
\alias{sumstat_seg_sites}
\title{Returns the Segregation Sites Statistics from simulations}
\usage{
sumstat_seg_sites(name = "seg_sites", transformation = identity)
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{transformation}{An optional function for transforming the results
of the statistic. If specified, the results of the transformation are
returned instead of the original values.}
}
\description{
Returns the Segregation Sites Statistics from simulations
}
\examples{
set.seed(50)
model <- coal_model(5, 1) + feat_mutation(5) + sumstat_seg_sites("segsites")
stats <- simulate(model)
print(stats$segsites)
}

