% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_sfs.R
\name{sumstat_sfs}
\alias{sumstat_sfs}
\title{Calculates the Site Frequency Spectrum from simulations}
\usage{
sumstat_sfs(name = "sfs", population = "all", transformation = identity)
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}

\item{population}{Either the number of a population for with the SFS is
calculated, or \code{"all"} to calculate the combined SFS of all
populations.}

\item{transformation}{An optional function for transforming the results
of the statistic. If specified, the results of the transformation are
returned instead of the original values.}
}
\description{
Calculates the Site Frequency Spectrum from simulations
}
\examples{
set.seed(50)
model <- coal_model(5, 2) + feat_mutation(5) + sumstat_sfs()
stats <- simulate(model)
print(stats$sfs)
}

