% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumstat_trees.R
\name{sumstat_trees}
\alias{sumstat_trees}
\title{Returns ancestral tress in NEWICK fromat from simulations}
\usage{
sumstat_trees(group = 0)
}
\arguments{
\item{group}{The locus group for which this summary statistic is reported.
The default of `0` corresponds to all groups.}
}
\description{
Returns ancestral tress in NEWICK fromat from simulations
}

