% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_trees.R
\name{sumstat_trees}
\alias{sumstat_trees}
\title{Summary Statistic: Ancestral Trees}
\usage{
sumstat_trees(name = "trees")
}
\arguments{
\item{name}{The name of the summary statistic. When simulating a model,
the value of the statistics are written to an entry of the returned list
with this name. Summary statistic names must be unique in a model.}
}
\description{
This statistic returns ancestral tress in NEWICK format.
}
\examples{
# Without recombination:
model <- coal_model(4, 2) + sumstat_trees()
stats <- simulate(model)
print(stats$trees)

# With recombination:
model <- model + feat_recombination(5)
stats <- simulate(model)
print(stats$trees)
}

